Change type of ID to int64 for matching the Long type in JAVA

+++ daemon/persistence/auditable.go
@@ -4,13 +4,13 @@ import "time"
 
 //Auditable ...Map to table name auditables
 type Auditable struct {
-	ID            uint      `db:"id"`
+	ID            int64      `db:"id"`
 	CreatedAt     time.Time `db:"created_at"`
 	UpdatedAt     time.Time `db:"updated_at"`
 	Activity      string    `db:"activity"`
 	Originator    string    `db:"originator"`
-	ProjectID     uint      `db:"project_id"`
+	ProjectID     int64      `db:"project_id"`
 	ActivityTime  time.Time `db:"activity_time"`
-	Action        uint      `db:"action"`
-	ProjectItemID uint      `db:"project_item_id"`
+	Action        int64      `db:"action"`
+	ProjectItemID int64      `db:"project_item_id"`
 }
